# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Farmacia",
        "description": "Esta API se encarga de manejar la información de la Gestión de las Farmacias especificando cada funcionalidad con la que cuenta como la especificación de la misma Entidad",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/v1/farmacias": {
            "get": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Mostrar Farmacias",
                "operationId": "cd6133cfbeec708d37c78a777f4e651d",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Crear Farmacia",
                "operationId": "6065180595c26b1d927286c2d8cda7bd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "dirección": {
                                                "type": "string"
                                            },
                                            "latitud": {
                                                "type": "number"
                                            },
                                            "longitud": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "",
                                    "dirección": "",
                                    "latitud": null,
                                    "longitud": null
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Botica MBF Pharma"
                                        },
                                        "dirección": {
                                            "type": "string",
                                            "example": " AV. CESAR CANEVARO MZ. Q - 2, LT. 4 A.H. TREBOL AZUL, Av. César Canevaro, San Juan de Miraflores"
                                        },
                                        "latitud": {
                                            "type": "number",
                                            "example": -12.1945699
                                        },
                                        "longitud": {
                                            "type": "number",
                                            "example": -76.966001
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farmacia/{id}": {
            "get": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Obtener Farmacia",
                "operationId": "20a6e77afc4372d758628c2ef61ae940",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Botica MBF Pharma"
                                        },
                                        "dirección": {
                                            "type": "string",
                                            "example": " AV. CESAR CANEVARO MZ. Q - 2, LT. 4 A.H. TREBOL AZUL, Av. César Canevaro, San Juan de Miraflores"
                                        },
                                        "latitud": {
                                            "type": "number",
                                            "example": -12.1945699
                                        },
                                        "longitud": {
                                            "type": "number",
                                            "example": -76.966001
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farmacias/{id}": {
            "put": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Modificar Farmacia",
                "operationId": "02f652eac0136c094263f5f8536c2a39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "dirección": {
                                                "type": "string"
                                            },
                                            "latitud": {
                                                "type": "number"
                                            },
                                            "longitud": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "",
                                    "dirección": "",
                                    "latitud": null,
                                    "longitud": null
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Botica MBF Pharma"
                                        },
                                        "dirección": {
                                            "type": "string",
                                            "example": " AV. CESAR CANEVARO MZ. Q - 2, LT. 4 A.H. TREBOL AZUL, Av. César Canevaro, San Juan de Miraflores"
                                        },
                                        "latitud": {
                                            "type": "number",
                                            "example": -12.1945699
                                        },
                                        "longitud": {
                                            "type": "number",
                                            "example": -76.966001
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Farmacia"
                ],
                "summary": "Borrar Farmacia",
                "operationId": "635d71332a9c396a4311e8e46c33ab69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Farmacia borrada satisfactoriamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "securityScheme": "apiAuth"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer",
            "securityScheme": "apiAuth"
        }
    }
}